pyproject.toml
[project]
name = "abelohost"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "black>=26.1.0",
    "coverage>=7.13.4",
    "fastapi>=0.128.5",
    "httpx>=0.28.1",
    "pandas>=3.0.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-dotenv>=0.5.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
